For #42679 #42966 and #43224, config descriptor fixes and offline mode.
Configuration descriptors
-------------------------
The configuration descriptors can now be used to describe both descriptor-based pipeline configurations and path-based pipeline configurations. This allows code running on top of Toolkit to see both worlds as a single environment, each with its own quirk hidden behind a clean interface. We're also adding support for python interpreter files inside the configuration so any project using the configuration will now be able to use a custom interpreter.

Offline mode
------------
Toolkit now allows to share a bundle cache with another process, which means that it is now possible to pre-package all the bundles at a centralized location and reuse it across processes, for example in the Shotgun Desktop. For example, then Shotgun Desktop can now be package with a complete configuration and used without an Internet connection.

Closes https://github.com/shotgunsoftware/tk-core/pull/475
Closes https://github.com/shotgunsoftware/tk-core/pull/479