Ticket/43145 add session refresh loop in desktop (#487)
* For #43145: modified usage of saml_expiration to help as a indication if sso is being used or not

* For #43145: changes from code review

* [minor] Added std header and docs to appveyor CI file.

* [minor] Added badge for Linting CI

Upgraded travis badge to be retina svg. Minor formatting tweaks so it takes up a little less space on the page.

* [minor] Improved resilience for failed app store downloads (#490)

Adds better protection from a case when the downloads succeed but the writing of event log entries to the app store fail. This is an unusual case but has been encountered in the wild.

* [minor] Handles non-ascii chars in entity names when building tooltips.

* [minor] Commenting out two tests that silently fail.
These tests will need to be revisited.

* For 41118. [minor] Adding debug messages to bootstrap backup cleanups for easier QA-ing of issues with the log file (#493)

For 41118. [minor] Adding debug messages to bootstrap backup cleanups for easier QA-ing of issues with the log file (#493)

* [minor] Upgrades all unit test third parties to their latest versions.
This includes the latest versions of
- mock
- unittests
- coverage

Also, a new script has been introduced at tests/python/upgrade_third_party.py that allows to upgrade our third parties. The list of third parties to install are picked from a requirements files that lives alongside it.

* [minor] Adding tests/coverage_html_report to .gitignore

* For #41415: Updates the mockgun schema. (#494)

* Updates the mockgun schema, requiring removal of legacy entity types from various tests.

* For #42679 #42966 and #43224, config descriptor fixes and offline mode.

Configuration descriptors
-------------------------
The configuration descriptors can now be used to describe both descriptor-based pipeline configurations and path-based pipeline configurations. This allows code running on top of Toolkit to see both worlds as a single environment, each with its own quirk hidden behind a clean interface. We're also adding support for python interpreter files inside the configuration so any project using the configuration will now be able to use a custom interpreter.

Offline mode
------------
Toolkit now allows to share a bundle cache with another process, which means that it is now possible to pre-package all the bundles at a centralized location and reuse it across processes, for example in the Shotgun Desktop. For example, then Shotgun Desktop can now be package with a complete configuration and used without an Internet connection.

Closes https://github.com/shotgunsoftware/tk-core/pull/475
Closes https://github.com/shotgunsoftware/tk-core/pull/479

* For #43145: implemented changes and cleanup required for the code reviewers

* For #43145: Removed parameter saml_expiration from multipile method signatures, as it can be obtained from the cookies

* For #43145: Added import of ShotgunSamlUser in the init of module authentication

* For #43145: Changes following code review

* Addendum for #42679, can't jump into more recent more.
Configurations now have transaction markers that indicate when an update started and ended. If the ended marker is missing, the configuration will be updated again since it is in an incoherent state.
There was also a performance issue with resolve_descriptor which wrote the configuration to disk when it wasn't needed.

* Revert "Merge branch 'master' into ticket/43145_add_session_refresh_loop_in_desktop"

This reverts commit ae1e150c066832f2bf102709c3382b69d7213dff, reversing
changes made to 16022e0a0e03fb143bef93538d642155c082f7ae.

* For #43145: Changes following code review from Manne

* For #43145: More changes following code review from Manne