for #34908: Expand env vars in interpreter files on Windows (#504)
We have functionality to expand env vars in interpreter files on Mac/Linux but this was previously not ported to Windows. This adds that support so that users can specify an environment variable as part of their interpreter settings on Windows and it gets correctly picked up by the tank.bat command.