For #41460, improved push_config.
Allowed the "push_config" command to be run from the API and added unit tests for it.
Fixed an edge case where the same backup target could be used if multiple push to the same target were run within a second (mainly for unit testing).
Reformatted push_pc.py code to be as much as possible on the left (which is the reason why we avoid to have long lines).
Issue an error if the symlink option is used on a platform not supporting it (Windows).
Improved comments in a couple of places.
Revisited backup "move": try to use `os.rename` first (which is what `shutil.move` does), fallback on `shutil.copytree` and then `shutil.rmtree` so there is always a valid config folder available.
Ensure the user always knows that a backup is available if it is the case, even in case of errors.
Added  `filesystem.get_unused_path` helper and added unit tests for it.
Added a note to `filesystem.move_folder` mentioning that the directory being moved is *not* removed, which is different from the expected behavior.