For #41460, fixes.
Added missing unit test file.
Fixed push_config test for Windows.
Fixed error message mentioning symlinks are *not* supported.
Used `filesystem.safe_delete_folder` which is more robust than `shutil.rmtree`.
Flag that a backup is available as soon as this is known, so errors happening later don't prevent this to be mentioned to the user.