For #44082: Keeps track of the source entity as given to context factory functions. (#513)
Adds a source_entity property to Context. This will contain a reference to the entity dictionary that was used to construct the context. This might be different than the context's "entity" property, as some entity types (most notably PublishedFile) have a context built from them that represents what they're linked to. In this case, the context's entity will likely be a Shot or Asset, and its source_entity will be the PublishedFile.

A specific use case for the above is in a pick_environment hook. When there's a desire to have an environment in a config specific to PublishedFile entities, the pick_environment hook can check the given context's source_entity to know that it needs to return that environment.