For #43063: Defaults the documentation url to the user guide.
Fixes an issue where clicking the 'Documentation' button in an app's info sidebar does not perform any action unless set in the app's info.yml.
By having the 'documentation_url' property on the Descriptor default to toolkit's user guide (instead of None), the user will be directed to this webpage if no specific entry is made in info.yml. Closes #515.