For #44807: Path cache optimizations (#532)
This implements two optimizations for path cache:

* Inner loop for secondary entities:

This removes inner loop scenarios when syncing and can drastically improve path cache sync performance in certain scenarios.

When secondary entities are used, an additional check was carried out to ensure that a path wasn't already registered in the db. This was done in a non-optimal way, meaning that as the number of records grew in the db (especially for step), syncning could blow up in to a hidden o(n^2) complexity type scenario, causing millions of operations to be carried out.

This kills the complexity and issues a single query to the db for every record tested.

* Batching of find queries:

Find queries are batched into smaller chunks in order to help the shotgun backend db.