For #44807: Fix for getting entities from an empty list of folder ids.
In PathCache, when getting a list of FilesystemLocation entities from a list of folder_ids, we now properly return an empty list when given an empty list of folder_ids. Prior to this fix, the result was that ALL FilesystemLocation entities for the project were queried and returned, which was not the correct behavior. That result should have been limited to situations where a None was passed through the folder_ids argument.