For #44135: Properly handles multi-tiered include structures in configs. (#528)
In a config structure similar to that of tk-config-basic, the environment API needs to be able to properly resolve what yml file a bundle's location descriptor resides in before updating it. This did not work properly previously, and the result was that included locations would often be flattened: the included value would be flattened. This did not break the project, but it did destroy the include structure of the config. Now, running a tank updates on a config will always properly resolve included bundle locations and ensure that the right yml files are updated while preserving the config's include structure.