#44757: Environment variable to specify the bundle cache path. (#529)
Adds support for the SHOTGUN_BUNDLE_CACHE_PATH environment variable that allows one to specify the location of the path on disk (possibly shared) to which all bundles would be downloaded to and made available from.

To unify the process of downloading descriptors locally, it introduces the IODescriptorDownloadable type. By first downloading a descriptor to a temporary path and then moving it to the intended bundle cache path atomically, it guards against multiple process attempting to download it simultaneously to a shared bundle cache path.