For # 45071: Fixes reference to self in a static function. (#536)
In bundle.py, we have logic for resolving template-type settings for bundles that can have a hook defined as their value. In this situation, the referenced core hook is called, which will then return the value for the setting.

During the major refactoring work that came about as part of the zero config project, one of the bundle instance methods was converted to be a static function. The regression occurs as a result of a lingering reference to "self" in the now-static function. The fix is to default to the current bundle, but offer the option of passing in a bundle object explicitly.