[minor] Removes fallback to current_bundle in bundle.py.
This was not possible in some cases depending on when the function to resolve bundle settings is called. The None bundle is handled gracefully, so the extra step of looking up the current bundle is not required anyway. The times when a bundle is needed are limited to only when core hooks are used to resolve template-type bundle settings, in which case the call to resolve the setting will have come from the bundle itself.