For #45071: [minor] Fix for zero-config cases.
Situations where a bundle isn't provided to the bundle settings resolution function occur in a zero config context. The fallback behavior here, previously, was to pull the current bundle from sgtk.platform.current_bundle. There are cases where this is not possible, however, and the resulting exceptions were not handled gracefully.

Instead, we now pass through the default None value for the bundle when resolving settings deeper in the API. This case is handled gracefully, so no ill effects are seen. This is because the only case where the bundle is explicitly needed is when resolving template-type bundle settings that are making use of a core hook to resolve the value. This case only arises when a bundle is resolving its own settings when it is initialized, in which case the bundle will pass itself to the resolution function.