#45817 Removes settings validation from sw launch (#554)
Fixes an issue with tk-maya when creating SoftwareLauncher instances if there is no filesystem structure on disk for the context. The engine's template_project setting will not validate preventing the launcher from being created. This fix simply omits the validation. We are scheduling a fix which will replace this with partial validation logic in an upcoming release.