#45876 Safe copy fallback for bundle cache I/O transactions (#559)
Descriptors are downloaded into a bundle_cache/tmp folder and then moved into their final location using a rename operation. Renames can fail (seen in the wild on windows 8.1) where file locking may prevent the files to be removed out of the original temp location. This change adds a fallback path where if the rename fails, a copy-then-graceful-delete approach is attempted.

If this in turn fails and in case of concurrent access, a transaction marker has been added to ensure stability. A `tk-metadata` folder is written into the bundle cache and once the content has been moved into place, a `install_complete` file is written into this folder.

This also adds a `--no-hardlinks` option to all git clone commands, ensuring that no hard links are created when handling local repos. This ensures better i/o integrity for all repo operations and avoid python raising 'file is the same' exceptions when trying to copy files.

Adds a `--debug` flag to the test runner.