For #40825, bootstrap from cache This removes the need to copy the content of a configuration into the unmanaged configuration when there is an update.
The PipelineConfiguration class was refactored so it could use the descriptor member to locate the configuration bundle on disk.

This also introduces a `features` API that allows the bootstrap process to introspect the core it is about to install and switch to without having to actually import the module. This is done with the helpof the `info.yml`. We're not marking the API as public for now and since it's only being used by the bootstrapper at the moment we don't have to worry about breaking changes to the API if we decide to tweak it.
Closes https://github.com/shotgunsoftware/tk-core/pull/565