Feature/sso (#572)
* Adds a missing import of sgtk in bundle.py.

* [minor] Removes fallback to current_bundle in bundle.py.

This was not possible in some cases depending on when the function to resolve bundle settings is called. The None bundle is handled gracefully, so the extra step of looking up the current bundle is not required anyway. The times when a bundle is needed are limited to only when core hooks are used to resolve template-type bundle settings, in which case the call to resolve the setting will have come from the bundle itself.

* Initial implementation of SSO support in tk-core.

Forget to address one of the PR comments

Hiding of unsupported browser message + fixed implementation of are_credentials_expired in SSO context.

Added a destructor just to log the saml2_sso object destruction

First set of changes requested from the code review by Manne and JF

Include fixes for bugs #46059 and #46071

More changes following the code review