Allow updating arbitrary env dictionaries
Allows the user to provide an alternate dictionary to update instead of just using os.environ. The enables users to create or modify environment dictionaries other than the active one. So a user could, for example, copy the current environment, modify the copy, and then provide then supply that copy to any potential consumer without modifying the current runtime.