For #40825, run a cached configuration's folder from the bundle cache. This branches removes the need to copy a configuration from the bundle cache into the cached configuration folder. Note that the core is still copied however.
This has two major benefits.

First, when running a `dev` descriptor for a config, a cached configuration will be able to pick up changes to the configuration after a reload.
Second, it speeds up the bootstrap time by removing the need to copy all the environment files and hooks into the cached configuration.

Closes https://github.com/shotgunsoftware/tk-core/pull/565