#46697 Fix for PyQt4 with authentication dialog
This is to fix an issue where the login screen would error when being created in an environment running PyQt4.

The fix is to just explicitly state that we want to create a QtCore.QUrl when passing the URL through to the setUserStyleSheetUrl method. Without this, on PyQt4 it would complain that it got the wrong argument type. closes #606