For #46874: Add templates_from_path function to return multiple templates from a path (#611)
In cases where you have clashing templates, specifically string templates, you may wish to collect all the matching templates to then further filter by template name (eg: if 'shot' in template_name, or if 'asset' in template_name). This opens up context-specific template searches, or the ability to identify the context of a file simply from a filename, rather than a full path. 
This update simply abstracts the template search part of template_from_path to allow users to obtain all matching templates from a path, not just one, and more importantly, it does not raise an error if multiple paths are found.