#46650 Adds descriptor uri to return value (#632)
Adds a descriptor_source_uri return value to the ToolkitManager.get_pipeline_configurations() method. This is to better support the case where a user has defined a descriptor (e.g. sgtk:descriptor:app_store?name=tk-config-basic) in a pipeline config and we want to be able to resolve the fact that this is a config uri which is implicitly tracking the latest version.