SG-4682 [minor] Tweaks to fix Open Sans style on Windows 10 in PySide2. (#645)
This contains a minor tweak/hack to make Windows 10 pick the correct font style when font-weight is used in conjunction with Open Sans. We're installing the ttfs for Open Sans in reverse alphanumeric order now, so that the "light" style font is installed before the "condensed light" style. Windows seems to prefer what's installed first when it's determining style based on font-weight settings in qss.