SG-7596, descriptor download from hook (#655)
Introduces a new core level hook named bootstrap.py. For now, it can be used to override the download of bundles by taking over the can_cache_bundle and populate_bundle_cache_entry methods.

As an example, this allows a user to cache certain bundles inside Shotgun and have them be downloaded from Shotgun instead of their original source. This is useful when dealing when a computer can't access the internet to download bundles from the Toolkit AppStore. This can also be used to distribute application code to users that do not have access to the git repositories that track your application code.