For SG-8449, path cache sandboxing fix (#667)
Path caches for descriptor-based configurations will now NOT use the plugin id in the folder suffix for the path cache.

This will allow reuse of the path cache between DCCs that use the same pipeline configuration and project, requiring less syncs of the path cache. It also fixes an issue when launching apps in an asset or task context from Shotgun.