[minor] Coverage speedup on CI (#678)
The version of coverage installed in the repo was pip installed on macOS. Given that the coverage module has a fast tracer that is a compiled module and a slower tracer that is pure Python, it meant that the repo had a macOS binary for tracing. Considering we are running CI on Linux and soon Windows, this means that the tests are running much slower than they could.

To resolve this problem, we will not have a copy of coverage inside the repo anymore and require users to have installed it themselves. This means that CI is going to run its own copy, which is going to have the binary tracer.