SG-17968 Implements a look up for previously processed descriptors. (#776)
This provides a performance increase when running updates on a config. I now doesn't look up update each time if an identical descriptor was previously processed. This should provide performance increases to both default and default2 styled configs.

Mostly the benefit will be to default2 styled configs, where all app, engine and framework locations are stored in one file and pointed at by the various instances of their settings. This means that once a descriptor has been processed once, the next time it is found it won't try to look for updates again.

It also ensures that the test cover the changes made and the the end result of an update stays the same.