Fixes the integration tests by using a simple test config.
app store and git setup project tests now use a new dedicated simple testing config,
which means testing time is a lot faster due to less dependancies.

Also removes a skip, now that tk-shell is Python 3 compatible.

Updates the test config components, but not to the latest so that the update test can still run.